refactor(nav): rename the Server page to Storage, scope the update button
BEHAVIOUR CHANGE.

The page only ever showed disk and database usage, so it is now /storage.
/server redirects, so existing links and bookmarks still land.

The update-status button was shown on every route except SiteDetail. It is
now opt-in per route via `meta.showUpdateStatus`, enabled on Sites,
Marketplace and Updates. Its poll therefore stops on pages that never
showed it. The mobile and desktop shells had drifted apart on this; both
now read the same flag.
