+++ admin/frontend/dashboard/src/components/navigation/list.ts
@@ -2,7 +2,7 @@ export const sidebarSections = [
   {
     items: [
       { label: 'Sites', icon: 'lucide-globe', to: '/sites' },
-      { label: 'Server', icon: 'lucide-server', to: '/server' },
+      { label: 'Storage', icon: 'lucide-hard-drive', to: '/storage' },
       { label: 'Marketplace', icon: 'lucide-store', to: '/marketplace' },
     ],
   },
